14 files updated, please read:
I added many new libraries to our settings that we need for boto3 as well as users. So I added boto3 library storages, which is aws third party to conenct our s3 bucket. So I created new views, and updated the entire upload_journal view,html,js so we can store our uplaods in our S3. Also documents are now grabbed straight from the S3 instead of a hardcoded presigned url, where continue reading now generates a presigned-url and redirects to a new window. As well as changed the entire views and Model to use CustomUser (or abstract user) which allows us to customize certain metadata pieces by still using built in django libraries and functions. This will allow us to cutstomize the student and editor dashboard a lot easier. Users will be able to save documents and cann be displayed in the dashboard. Also was working on adding a "Cite" view and the "Save Document" view, didnt get to. These changes are imperative to change asap so eveyrones will be compatible.