15 files changed:
The entire backend is changed, so it might create problems with the merge. I learned after more researching that I was double creating backend instances and data that was not needed. SO i was hardwriting all the SQL entries when it was also being created in migrations. So migrations is what creates all the backend data in mysql. So the backend is cahnged to function more with the database and frontend queries. Created more CustomUser and Document data to be used, seed_documents is created to populate the documents now instead of the sql files. Users now have different permissions, Homepage buttons (Cite Download and Save) work now for their functionalities. Changed the S3 bcuket to be more organized for storage usage (image/ and journal/) for when documents are uploaded. And hopw the file path's are made. Added a lot that I still cant remember.